requirements.txt
idem>=5
libvirt-python
